refactor: 상수 제거 및 함수 추출
- 상수 ROUTES 제거
- 모델 체크하는 isCheckModel 함수
- isCheckModel 함수를 util로 옮기면서 파일명 변경 (api-key-check.ts > check.ts)
