+++ src/constant/index.ts
@@ -1,9 +1,3 @@
-export const ROUTES = {
-  MAIN: "/",
-  API_KEY_SETTING: "/setting/api-key-setting",
-  IMAGE_GALLERY: "/setting/image-gallery",
-};
-
 export const MENU_LIST = [
   { id: "1", value: "profile", name: "Profile" },
   { id: "2", value: "api-key-setting", name: "API Key Setting" },
@@ -14,7 +8,6 @@ export const MODELS = {
   "GPT-3.5": "gpt-3.5-turbo",
   "GPT-4": "gpt-4-turbo-preview",
   "DALL-E-3": "dall-e-3",
-  BING: "bing",
   LLAMA2: "llama2-70b-4096",
   MIXTRAL: "mixtral-8x7b-32768",
   GEMINI: "gemini-pro",
